refactor: separate domain package (#37)
