--- api/src/main/kotlin/team/backend/config/StaticRoutingConfiguration.kt
@@ -1,13 +1,0 @@
-package team.backend.config
-
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.reactive.config.ResourceHandlerRegistry
-import org.springframework.web.reactive.config.WebFluxConfigurer
-
-@Configuration
-class StaticRoutingConfiguration: WebFluxConfigurer {
-    override fun addResourceHandlers(registry: ResourceHandlerRegistry) {
-        registry.addResourceHandler("/static/**").addResourceLocations("classpath:/static/")
-        registry.addResourceHandler("swagger-ui.html").addResourceLocations("classpath:/static/swagger-ui/")
-    }
-}
