--- api/src/main/kotlin/team/backend/domain/entity/Member.kt
@@ -1,28 +1,0 @@
-package team.backend.domain.entity
-
-import team.backend.infrastructure.BaseEntity
-import jakarta.persistence.Column
-import jakarta.persistence.Entity
-import jakarta.persistence.FetchType
-import jakarta.persistence.GeneratedValue
-import jakarta.persistence.GenerationType
-import jakarta.persistence.Id
-import jakarta.persistence.JoinColumn
-import jakarta.persistence.ManyToOne
-import jakarta.persistence.Table
-
-@Entity
-@Table(name = "tbl_member")
-class Member(
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "member_id")
-    val id: Long? = null,
-
-    @Column(name = "member_name", nullable = false)
-    val name: String,
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "job_id", nullable = false)
-    val job: Job,
-): BaseEntity()
