--- api/src/main/kotlin/team/backend/domain/repository/MemberQuizAnswerStore.kt
@@ -1,8 +1,0 @@
-package team.backend.domain.repository
-
-import team.backend.domain.entity.MemberQuizAnswer
-
-interface MemberQuizAnswerStore {
-
-    suspend fun save(entity: MemberQuizAnswer)
-}
