--- api/src/main/kotlin/team/backend/domain/repository/MemberReader.kt
@@ -1,8 +1,0 @@
-package team.backend.domain.repository
-
-import team.backend.domain.entity.Member
-
-interface MemberReader {
-
-    suspend fun get(id: Long): Member
-}
