--- api/src/main/kotlin/team/backend/domain/repository/MemberStore.kt
@@ -1,4 +1,0 @@
-package team.backend.domain.repository
-
-interface MemberStore {
-}
