--- api/src/main/kotlin/team/backend/domain/repository/PromptReader.kt
@@ -1,10 +1,0 @@
-package team.backend.domain.repository
-
-import team.backend.domain.entity.PositionType
-import team.backend.domain.entity.Prompt
-import team.backend.domain.entity.PromptType
-
-interface PromptReader {
-
-    suspend fun get(promptType: PromptType, positionType: PositionType): Prompt
-}
