--- api/src/main/kotlin/team/backend/domain/repository/PromptStore.kt
@@ -1,4 +1,0 @@
-package team.backend.domain.repository
-
-interface PromptStore {
-}
