--- api/src/main/kotlin/team/backend/domain/repository/QuizReader.kt
@@ -1,8 +1,0 @@
-package team.backend.domain.repository
-
-import team.backend.domain.entity.Quiz
-
-interface QuizReader {
-
-    suspend fun get(id: Long): Quiz
-}
