--- api/src/main/kotlin/team/backend/domain/repository/QuizStore.kt
@@ -1,4 +1,0 @@
-package team.backend.domain.repository
-
-interface QuizStore {
-}
