--- api/src/main/kotlin/team/backend/domain/service/QuizService.kt
@@ -1,8 +1,0 @@
-package team.backend.domain.service
-
-import team.backend.domain.dto.command.QuizCommand
-
-interface QuizService {
-
-    suspend fun answer(command: QuizCommand.AnswerRequest): QuizCommand.AnswerResponse
-}
