+++ api/src/main/kotlin/team/backend/prompt/domain/repository/PromptReader.kt
@@ -1,0 +1,10 @@
+package team.backend.prompt.domain.repository
+
+import team.backend.prompt.domain.entity.PositionType
+import team.backend.prompt.domain.entity.Prompt
+import team.backend.prompt.domain.entity.PromptType
+
+interface PromptReader {
+
+    suspend fun get(promptType: PromptType, positionType: PositionType): Prompt
+}
