+++ api/src/main/kotlin/team/backend/prompt/domain/repository/PromptStore.kt
@@ -1,0 +1,4 @@
+package team.backend.prompt.domain.repository
+
+interface PromptStore {
+}
