+++ api/src/main/kotlin/team/backend/quiz/domain/repository/QuizStore.kt
@@ -1,0 +1,4 @@
+package team.backend.quiz.domain.repository
+
+interface QuizStore {
+}
