+++ api/src/main/kotlin/team/backend/quiz/domain/service/QuizService.kt
@@ -1,0 +1,8 @@
+package team.backend.quiz.domain.service
+
+import team.backend.quiz.domain.dto.command.QuizCommand
+
+interface QuizService {
+
+    suspend fun answer(command: QuizCommand.AnswerRequest): QuizCommand.AnswerResponse
+}
